Remove panics from handlePtraceSyscallRequest and attach.
A lot of these panics are valid -- but in an effort to consolidate
systrap panics to a fewer places so that some can be localized to
affect individual subprocesses rather than entire sandboxes --
we should convert them to errors.

Note syscallThread.detach still panics after this CL. It is not as
trivial to remove those due to us already being attached.

PiperOrigin-RevId: 603802250